requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==3.0.1
django-crispy-forms==1.7.2
django-extensions==2.1.9
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-taggit==1.1.0
djangorestframework==3.9.4
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
markdown2==2.3.7
MarkupSafe==1.1.1
numpy==1.16.4
oauthlib==3.0.1
pandas==0.24.2
Pillow==6.0.0
post==2019.4.13
psycopg2==2.7.7
public==2019.4.13
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.4
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
soupsieve==1.9.2
urllib3==1.25.3
virtualenv==16.4.0
Werkzeug==0.15.4
whitenoise==4.1.2
